edit dialog, delete lock action, improved date picker focus

+++ src/components/App/App.js
@@ -14,15 +14,29 @@ export default Vue.component('App', {
     data() {
         return {
             tabActive: 0,
+
+            timeDeleteUnlock: null,
+
+            isEditDialogOpen: false, 
+
             timeStart: "",
             timeEnd: "",
             label: "",
-            date: ""
+            date: "",
+
+            edit: {
+                id: "",
+                timeStart: "",
+                timeEnd: "",
+                label: "",
+                date: "",
+            }
         }
     },
     computed: {
         ...mapGetters({
-            records: 'getRecordsCollection'
+            records: 'getRecordsCollection',
+            recordById: 'getRecordById'
         }),
 
         isInstallReady() {
@@ -44,6 +58,7 @@ export default Vue.component('App', {
         ...mapActions({
             retrieveSavedRecords: 'retrieveSavedRecords',
             addRecord: 'addRecord',
+            editRecord: 'editRecord',
             deleteRecord: 'deleteRecord'
         }),
 
@@ -73,7 +88,31 @@ export default Vue.component('App', {
             this.date = "";
         },
 
+        requestDeleteUnlock() {
+            let isActionAccepted = true;
+            let secondsFromTheUnlock = dateFns.differenceInSeconds(new Date(), this.timeDeleteUnlock);
+            let isUnlockTimeExpired = secondsFromTheUnlock > 5;
+
+            // isMoreThanUnlockTime
+            if (isUnlockTimeExpired) {
+                isActionAccepted = confirm("Are you sure you want to delete this record?");
+            }
+
+            // isLessThanUnlockTime & confirm
+            if (!isUnlockTimeExpired || isActionAccepted) {
+                this.timeDeleteUnlock = new Date();
+            }
+
+            return !isActionAccepted;
+        },
+
         deleteLog(id) {
+            let isLocked = this.requestDeleteUnlock();
+
+            if (isLocked) {
+                return;
+            }
+
             this.deleteRecord(id);
         },
 
@@ -88,6 +127,46 @@ export default Vue.component('App', {
             this.label = label;
 
             this.onSubmit();
+        },
+
+        resetEditLog() {
+            this.isEditDialogOpen = false;
+            this.edit = {
+                id: "",
+                timeStart: "",
+                timeEnd: "",
+                label: "",
+                date: "",
+            };
+        },
+
+        openEditDialog(id) {
+            this.isEditDialogOpen = true;
+
+            let {
+                timeStart,
+                timeEnd,
+                label,
+                date
+            } = this.recordById(id);
+
+            this.edit = {
+                id,
+                timeStart,
+                timeEnd,
+                label,
+                date,
+            };
+        },
+
+        closeEditDialog() {
+            this.resetEditLog();
+        },
+
+        saveEditDialog() {
+            this.editRecord(this.edit);
+
+            this.resetEditLog();
         }
     },
     template: `
@@ -141,8 +220,8 @@ export default Vue.component('App', {
 
             <h2 class="App_title">Records</h2>
 
-            <ul class="App_list" v-for="record in records">
-                <li>
+            <ul class="App_list">
+                <li v-for="record in records">
                     <mdc-card>
                         <mdc-card-header
                             :title="record.date + ' ' + record.label">
@@ -154,6 +233,11 @@ export default Vue.component('App', {
                             <mdc-card-action-icons>
 
                                 <mdc-card-action-icon 
+                                    icon="edit" 
+                                    @click="openEditDialog(record.id)"
+                                />
+
+                                <mdc-card-action-icon 
                                     icon="delete" 
                                     @click="deleteLog(record.id)"
                                 />
@@ -163,6 +247,25 @@ export default Vue.component('App', {
                     </mdc-card>
                 </li>
             </ul>
+            
+            <mdc-dialog 
+                v-model="isEditDialogOpen"
+                title="Edit" 
+                accept="Save" 
+                cancel="Decline"
+                @accept="saveEditDialog" 
+                @cancel="closeEditDialog">
+                
+                <div class="App_edit-modal-content-layout">
+                    <date-picker v-model="edit.date" :value="edit.date" label="date"/>
+
+                    <time-picker v-model="edit.timeStart" :value="edit.timeStart" type="time" label="time start"/>
+
+                    <time-picker v-model="edit.timeEnd" :value="edit.timeEnd" type="time" label="time end"/>
+
+                    <mdc-textfield v-model="edit.label" label="label" />
+                </div>
+            </mdc-dialog>
 
             <p v-if="!records.length">No items saved</p>
